Place code in ``code`` formatting.
With the previous version of the code, function signatures etc. get inserted as plain text.  A common python construct ``f(**kw)`` thus generates error since structured text thinks this is the start of **bold** text, but then complains because the start tag ``**`` is never closed.

The solution here is to place all code in double hash tags so it is formatted verbatim without interpreting characters as inline markup.

(Technically, this might only be supported by reStructuredText, but since this is the standard for python documentation, that is probably fine.