Added Travis, appveyor, tox, manylinux for 3.6 branch.
Copied from master branch, but removed the Python 3 stuff,
which is unsupported on the 3.6 branch. Well, 3.1 seems supported,
but that is very old, and on 3.3 I got errors locally.

Also removed 'docs' from the tox envs, as that proved tricky to get to work,
and has changed much.

Removed 'TODO_test_classProvides_fails_for_odd_class' test method,
which would always fail but be ignored due to its name,
except that the coverage tests would *not* ignore it, leading to a failure.