First pass at github actions.
Fixes #225. Someone will need to add a `TWINE_PASSWORD` GitHub repository secret that is
a token for `zope.eggbuilder` to upload to zope.interface.

Builds and uploads manylinux32/64/aarch64 wheels.

Builds and uploads Mac wheels.

Builds the docs.

Runs tests with the C extension and without the C extension.

Reports coverage to coveralls.

Has the start of an environment to do linting.

Removes .travis.yml